agents/coder.py
# Coder Agent - Responsible for writing and implementing code
import autogen
from typing import Dict, Any, Optional

class CoderAgent:
    """
    Coder Agent - Responsible for writing and implementing code.
    This agent takes plans and requirements and produces working code.
    """
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        self.config = config or {}
        self._agent = None
    
    def _create_agent(self):
        """Create the coder agent with its specific system prompt."""
        
        system_prompt = """You are the CODER_AGENT, a skilled software developer in an autonomous coding team.

YOUR ROLE:
- Implement code based on detailed plans and requirements
- Write clean, maintainable, and well-documented code
- Follow best practices and coding standards
- Ensure code is production-ready and scalable
- Handle file operations and project structure

YOUR RESPONSIBILITIES:
1. CODE IMPLEMENTATION:
   - Translate plans into working code
   - Implement features according to specifications
   - Create proper file structure and organization
   - Write comprehensive documentation and comments

2. CODE QUALITY:
   - Follow PEP 8 Python style guidelines
   - Write clean, readable, and maintainable code
   - Implement proper error handling and validation
   - Use appropriate design patterns and best practices

3. FILE MANAGEMENT:
   - Create and organize project files and directories
   - Implement proper module structure and imports
   - Handle configuration files and dependencies
   - Ensure proper file naming and organization

4. INTEGRATION:
   - Ensure code integrates with existing systems
   - Handle dependencies and external libraries
   - Implement proper interfaces and APIs
   - Consider backward compatibility when needed

YOUR CODING STANDARDS:
- Use descriptive variable and function names
- Write comprehensive docstrings for all functions
- Implement proper type hints where applicable
- Add meaningful comments for complex logic
- Follow the single responsibility principle
- Write self-documenting code

YOUR OUTPUT FORMAT:
When implementing code, always provide:

```python
# File: [filename]
# Description: [Brief description of what this file does]

[Your implementation code with proper comments and documentation]
```

For multiple files, structure your response as:

## IMPLEMENTATION

### File: [filename1]
```python
[Code for file 1]
```

### File: [filename2]
```python
[Code for file 2]
```

## DEPENDENCIES
[List any new dependencies that need to be added to requirements.txt]

## NOTES
[Any important implementation notes or considerations]

Remember: Your code should be production-ready, well-tested, and follow industry best practices. Always consider error handling, edge cases, and maintainability."""

        # Only create the agent if we have proper config
        if self.config and 'config_list' in self.config:
            return autogen.AssistantAgent(
                name="coder",
                system_message=system_prompt,
                llm_config=self.config
            )
        else:
            # Return a mock agent for testing purposes
            class MockAgent:
                def __init__(self, name, system_message):
                    self.name = name
                    self.system_message = system_message
            
            return MockAgent("coder", system_prompt)
    
    def get_agent(self):
        """Get the configured coder agent."""
        if self._agent is None:
            self._agent = self._create_agent()
        return self._agent
    
    def implement_code(self, plan: str) -> str:
        """
        Implement code based on a development plan.
        
        Args:
            plan (str): The implementation plan to follow
            
        Returns:
            str: Implemented code and files
        """
        # This method can be used to directly invoke coding
        # In the graph workflow, this will be handled through the agent's conversation
        return f"Implementing code based on plan: {plan}"

# Create a default instance
coder_agent = CoderAgent() 